Added 'support' for bi-directional inout pins
An inout pin is handled by splitting it into a seperate input and output pin, 
and fixing the net connectivity to match.

This operation is performed as a preprocessing step before the main vqm2blif 
conversion takes place, and uses the netlist generated by the VQM flex/bison 
parser. The related code in located in the preprocess.cpp and preprocess.h 
files.

The architecture file is now loaded directly after parsing the VQM.  This is
required to enable resolving the port direction of VQM primitives when 
splitting the inout pins (required to correctly adjust the nets attached
to inout pins).