Initial revision of code to handle inserting buffers to fix global net limitations in VPR.
This code checks the architecture file to find all blif .subckt
primitives that have pins marked with the is_non_clock_global='true'
attribute. 

Once these global pins are identified, it scans the netlist to
identify any (potentially) global nets that are connected to
a primitive global pin.

With the (potentially) global nets identified, it determines
each net's driver.  

If the driver is a global primitive pin, then the net is treated as
a global and any connection to a non-global pin has a global-to-local
(g2l) buffer inserted in the netlist.

If the driver is NOT a global primitive pin, but more than 50% of the
net's sinks are globals, the net is promoted to a global net, with a
local-to-global (l2g) buffer inserted at the driver's output.

If the driver is NOT a global primitive pin, and less than 50% of the
net's sinks are globals, the net is treated as a local net, and any
global sinks have a l2g buffer inserted before thier input.