Add new SDC parsing library and integrate into VPR.
The libsdc_parse library uses flex and bison to handle the 
actual SDC parsing. This enables:
    * Special cases such as escaped strings to work correctly
      (fixes issues with SDCs on some Titan23 benchmarks)
    * Command options to be re-ordered
    * Better (although still limited) support for nested commands
      like get_ports and get_clocks
    * Improved error checking and messages

Since libsdc_parse converts each SDC command into a struct,
it also provides a cleaner seperation between the parsing of
SDC commands and their interpretation in the tool.

Currently supports the same subset of SDC commands and options
as VPR's orginal SDC parser, but should be easier to extend
with additional features in the future.

The new parser has been verified to produce the same timing 
constraints in VPR (i.e. timing_constraints.echo) as the
original SDC parser on the following test cases:
    * vtr_flow/tasks/regression_tests/vtr_reg_nightly/vpr_reg_multiclock
    * vtr_flow/sdc/samples/*
    * The timing driven titan23 benchmarks


git-svn-id: https://vtr-verilog-to-routing.googlecode.com/svn/trunk@4243 8e3573b8-cf2c-4f14-ef6d-137439e28b8b