colouring of critical path highlighting updated.
New colouring explanation:
the head {clb,sub-block} is now a desaturated yellowish-green.
the driver {clb,sub-block} of the head is a light blue.
the rest of the tail is an even lighter blue.
the head {net,logical net} is a strong green.
the rest of the {nets,logical nets} are a grey-blue.

Also, the trailing logical nets are now coloured different, to be consisntent.

git-svn-id: https://vtr-verilog-to-routing.googlecode.com/svn/trunk@4349 8e3573b8-cf2c-4f14-ef6d-137439e28b8b