This commit has two main changes:
Change 1) We can now specify the parameters of an ipin cblock switch by point to a switch. Originally these parameters were specified in sizing/timing nodes like this:

    <sizing R_minW_nmos="8926" R_minW_pmos="16067" ipin_mux_trans_size="1.222260"/>
    <timing C_ipin_cblock="1.47e-15" T_ipin_cblock="7.247000e-11"/>

Now we can specify the ipin cblock  parameters as:

    <sizing R_minW_nmos="8926" R_minW_pmos="16067"/>
    <ipin_cblock switch="ipin_cblock_switch"/>

Where ipin_cblock_switch is the name of the switch. Either of these methods can be used.


Change 2) We can now specify multiple delays for a single switch as a function of the number of inputs. Normally a switch is specified like this:

    <switch type="mux" name="0" R="551" Cin=".77e-15" Cout="4e-15" Tdel="58e-12" mux_trans_size="2.630740" buf_size="27.645901"/>

Now we can specify multiple Tdel values as a function of fan-in like this:

    <switch type="mux" name="0" R="551" Cin=".77e-15" Cout="4e-15" mux_trans_size="2.630740" buf_size="27.645901">
      <Tdel num_inputs="2" delay="21e-12"/>
      <Tdel num_inputs="4" delay="30e-12"/>
      <Tdel num_inputs="8" delay="41e-12"/>
      <Tdel num_inputs="16" delay="58e-12"/>
      <Tdel num_inputs="32" delay="81e-12"/>
      <Tdel num_inputs="64" delay="113e-12"/>
    </switch>

We linearly extrapolate/interpolate to any missing points based on the two closest points. 
To allow a switch to have multiple fan-ins we create a new list of switches (g_rr_switch_inf) at the end of build_rr_graph based on the switch fan-ins determined from the rr graph. The rr nodes are then adjusted to point to the switches in this new list (instead of the list that contains switches from the architecture file).

git-svn-id: https://vtr-verilog-to-routing.googlecode.com/svn/trunk@4447 8e3573b8-cf2c-4f14-ef6d-137439e28b8b