Calling malloc_trim(0) after freeing switch block unordered map.
The switch block unordred_map can get quite large and it doesn't seem like the program is interested in releasing this memory back to the OS when the map is cleared. Calling malloc_trim forces the program to give unused heap space back to the OS which significantly reduces memory usage if multiple large benchmarks are run in parallel.



git-svn-id: https://vtr-verilog-to-routing.googlecode.com/svn/trunk@4677 8e3573b8-cf2c-4f14-ef6d-137439e28b8b