Fc can now be specified on a segment type basis.
The format for specifying Fc is largely unchanged but an optional override for segment types
can be specified. Example:

<!-- line specifying default fc values is unchanged -->
<fc default_in_type="frac" default_out_type="frac" default_in_val="0.15" default_out_val="0.1">
	<!-- Optional segment-based overrides -->
	<segment name="l4_global" in_val="0.5" out_val="0.4"/>
	<segment name="l16_global" in_val="0.3" out_val="0.3"/>
</fc>

The segment-based override specifies the name of the segment type (should match the names in 'segmentlist')
and the new input and output flexibilities. Note that input/output 'type' is not specified -- it is
inherited from the default_in/out_type. This may be changed in the future.

Also note that only pin-based *OR* segment-based Fc overrides can be specified for now, not both.