Reduce channel width for placement delay lookup.
When no channel width is specified, instead of picking a channel width
of 4x the maximum number of pins on a block, pick 4x the number of
pins on the FILL_TYPE (when doing the placement delay lookup we build
an FPGA of only IO and FILL_TYPE). This produces more reasonable channel
widths e.g. W=532 (instead of W=900) on the Titan Stratix IV
architecture capture.