Fix unused input counting in get_free_molecule_with_most_ext_inputs_for_cluster()
Previously the number of used (instead of unused) inputs were counted.
Since this function is only used as a fallback during packing (to pick
unrelated logic to fill-up a cluster) it has a relatively small QoR
impact.

Quick checks show there is no significant impact on Qor with the VTR
benchmarks, but there can be some reduction in DSP block usage on the
Titan benchmarks (e.g. 47 to 44 DSP blocks on des90).