Add support for more advanced multi-node <wireconn> specification
This allows multiple wire types with *different* switchpoints to be
specified together as the source switchpoints. For example:

<wireconn num_conns_type="to">
    <from type="L16" switchpoint="0,4,8,12"/>
    <from type="L4prime" switchpoint="0,1,2,3"/>
    <from type="L4" switchpoint="0,1,2,3"/>
    <to type="L4prime" switchpoint="0"/>
</wireconn>

Specifies that the from set is the union of all <from> tags enclosed
withing the <wireconn>, and similarity the <to> tags specify the
destination set.

Compared to the traditional 'inline' formulation (using the from_type,
from_switchpoint, to_type, to_switchpoint attributes) this allows
different switchpoints to be specified for each wiretype in the set.