Update RR graph generator to apply Fc to collections of pins
Previously Fc was applied on a per-pin basis. This could significant
deviation between the specified and actual Fc values (particularly at
low Fc for small collections of segments, where the number of segments
to be connected was < 0.5; such connections where then clipped to zero
and then bumped up to 2). We now specify Fc on collections of pins
(typically ports) which should help address this issue (since a larger
number of connections is calculated, this reduces the likely hood of
clipping to zero).

Also update Fc override handling in the architecture file. We now use
the <fc_override> tag, which replaces the previous <pin> and <segment>
overrides. The <fc_override> tag allows both pin and segment overrides
to be specified together (which was not previously supported).