Fix custom pin location specification for height > 1 blocks
Previously VPR erroneously rejected <loc> specifications with different
offsets.

This also relax restriction that custom pin locations can only be
specified on the perimeter of the block; since channels over-fly
blocks with height > 1, it is possible to connect to wires at
non-perimeter locations (e.g. at the horizontal mid-point of a
height > 1 block).