Add support for combined Range and Absolute difference QoR checks
The pass_requirements.txt format now supports the new directive:

    RangeAbs;<MinRatio>;<MaxRatio>;<AbsDiff>

Which specifies that a QoR metric must fall within the specified ratio
range, or absolute difference in order to pass. This is used for metrics
which may be small in absolute magnitude which can cause the ratio to vary
significantly despite the absolute difference being small (e.g. run-time
for small circuits).

For example:

    place_time;RangeAbs;0.10;10.0;2

specifies that the place time must be within 10x (relative) or 2 seconds
(absolute) of the golden value in order to pass.

Additionally the previous 'String' directive has been replaced by 'Equal'.

For example previously:

    vpr_status;String

indicated that an exact match was required. Since this can also be used
for numeric values, we now use the 'Equal' directive instead:

    vpr_status;Equal

which yields the same behaviour.

All relevant pass_requirement files have been updated to the new
formats.

parse_vtr_task.pl now also checks that specified directive is valid.