Allow either disconnected or constant connections to placement macro head blocks
Previously VPR relied on the head block of a carry chain macro to have a
disconnected cin pin. We now accept either a disconnected cin, or a cin
connected to a constant net.

Allowing constant nets fixes carry chain handling for the latest version
of the Titan benchmarks, which no longer strip out constant nets.