Treat constant nets a globals so they are not routed
Previously VPR treated the nets named 'vcc' or 'gnd' specially; they
were assumed to be constant and not routed. This approach was fragile
and did not detect nets constant nets with other names.

We now detect whether a net is constant from the netlist logic, and
do not route such nets.