Fix placement macro swapping on non-uniform device floorplans
Previously the macro swapping code assumed a uniform column based device
floorplan, and as a result only needed to check that a macro fit in the
'core' region of the device to ensure it is legal. [Since VPR is
currently only tested with vertical carry chain macros this check was
sufficient provided the 'core' region had columns of a uniform type
matching the macro].

Since we now support non-uniform floorplans the previous checks are no
longer sufficient. A macro is legal provided:
 * it fits anywhere on the device (i.e. there is no longer a distinction
   between the 'core' and 'perimeter'), and
 * the resources at the new location match those of the macro.

[Put another way, the previous swapping code assumed that macros were
translationally invariant, a condition which does not hold in general].

This commit adds checks to ensure that the resources requirements of a
macro are satisfied at the swapped-to location.

This fixes the placement of macros on non-uniform devices, and should
also support macros consisting of multiple block types.