Disable creation of switch blocks 'inside' width > 1 & height > 1 blocks
Previously when VPR always assumed it could create switch blocks
wherever channels crossed. Now that we support width > 1 blocks it is
possible for channels to cross *within* a large block.

This change updates the switch block creation code to detect when a
switch block would be created inside a block's boundary.

In the current implementation such switch blocks are skipped (not
created), but this could become user configurable in the future.