archfpga: Allow pin locations to be specified on both sides of a channel
Previously we treated the block sides (xoffset=0, yoffset=0, side=TOP)
and (xoffset=0, yoffset=1, side=BOTTOM) as equivalent. With the updates
to the RR graph this is no longer the case and these now specify either
the bottom or top of a horizontal routing channel respectively.

This commit removes the previous error checks which warned that the some
specifications were equivalent.