vpr: Don't connect perimeter block pins to non-adjacent channels
Blocks around the perimeter usually have their pins duplicated on all
sides, so that all the pins can connect to the single adjacent routing
channel. Due to bug all the pins were being connected (instead of only
those adjacent to the routing channel). This commit ensures that only
the pins adjacent to the routing channel are connected to it.