Move non-subtree libraries to libs/ folder
We have collected a large number of libaries in the VTR root directory.
To simplify the root directory common libraries will be moved under
libs/

Additionally, libaries which are developed externally (i.e. outside of
the VTR repository) will be moved under libs/EXTERNAL.

Currently libraries included using 'git subtree' are still in the root and will be moved next