vpr: Add is_configurable attribute to RR Graph edges
This attribute flags whether an edge in the RR graph is configurable by
the router (i.e. can be turned on/off).

Usually this attribute will be true, indicating that the edge
represents a configurable switch (e.g. which can be controlled via a bit
in the device configuration RAM).

However, in some instances the attribute will be false, indicating that
the edge is *not* backed by a configuration bit, and must be used if the
edge's source node is used.
This is used to model non-optional connections such as:
 * two wire segments which are electrically 'shorted' together
 * Non-tristate-able buffers between two wire segments
If these edges are encountered by the router, the associated sink nodes
*must* be expanded and added to the route tree to ensure accurate
capacitance/timing/wire usage information.

At this time the forced expansion is not yet implemented in the router.