vpr: Add initial recursive expansion of non-configurable edges (untested)
When a node is added to the heap, it's outgoing edges are explored and
any nodes connected via non-configurable edges are also added (recursively)
to the heap.

This should capture the 'forced' expansion nature of non-configurable
RR edges.

Note that this should ensure the router considers the forced expansion
nodes during it's exploration, however we currently do not handle them
when constructing the traceback/route tree (e.g. for delay calculation).