vpr: Fix traceback of nets with non-configurable RR edges
The previous implantation could add non-configurably connected nodes to
the traceback multiple times (once for each traceback update), causing
some routing resources to (incorrectly) appear to be overused.

This is fixed by considering *all* nodes current in the net's traceback
when the traceback is updated (instead of just those on the current
branch). This prevents an RR node from being added to the traceback
multiple times by different branches.