vpr: Add legality check for routing using non-configurable edges
This check ensures that any set of non-configurably connected nodes
or edges are *all* included in a net's routing, if *any* of the
non-configurably connected nodes or edges edges are used.
In other words, this check ensures that non-configurable edges are
correctly fully expanded in any net's routing.

The legality check has no effect of the architecture has no
non-configurable edges.