vpr: Ensure Netlist port's pins stored in ascending order of port index
Previously they were stored by order of creation, which for most netlist
results in the pins being added in the appropriate order. However, if
the netlist lists the ports out-of-order this would cause an error
(since we rely on this sorting for fast pin look-ups).

We now ensure the pin is added to the port in the correct location,
regardless of creation order.