vpr: Add support for reading extended BLIF (.eblif) circuit netlists
This adds support for the following BLIF extensions (pioneered by YOSYS):
    .conn   - Connects two nets together (analogous to verilog's 'assign' statement)
    .cname  - Specifies the name of an atom primitive (e.g. .names, .subckt)
    .param  - Specifies a param associated with an atom primitive (e.g. mode setting)
    .attr   - Specifies an attribute associated with an atom primitive (e.g. source file/line)

The current parser applies the .cname, .param, .attr to the last atom
primitive (.names, .latch, .subckt) which appeared in the netlist. For
example:

    .subckt dsp a=a b=b cin=cin cout=cout s=sum
    .cname my_adder
    .param mode adder
    .attr src my_verilog:32

would specify that the above 'dsp' .subckt would be named 'my_adder', that it
would have the parameter 'mode' set to 'adder', and the attribute
'src' would be set to 'my_verilog:32'.

The .conn statement allows direct connections between two wires. For
example:

    .model top
    .input a
    .output b

    .conn a b

    .end

specifies that 'a' and 'b' are direct connected together. In standard
BLIF this would require the insertion of .names buffers.

Currently VPR ignores any applied .param/.attr statements. In the future
these can be annotated on VPR's internal data structures and passed all
the way through. For now VPR prints a warning that the params/attrs are
ignored.