vpr: Add run-time --clustering_pin_feasibility_filter option
This option controls whether the clustering engine performs pin counting
to quickly filter infeasible molecules.

Provided the pin-counting code is always correct (i.e. it only returns
infeasible if the full router would return infeasible) this should only be a
run-time optimization with no impact on QoR.

However a run-time flag allows us to disable the pin-counting checks to
verify that the pin counter and full router agree, and provides a work
around if there are bugs in the pin counter (e.g. #278).