vpr: Add command-line option to control routing of constant nets
The new option '--constant_net_method' controls how nets driven by a
constant value are treated by VPR:
 * global: The nets are treated as global and not routed
 * route : The nets are treated as regular nets and routed

The default is 'global' to match previous behaviour.

Warnings are produced if a global net connects to a non-global pin.

In the longer term VPR should be taught to understand which pins in an
architecture are tie-able to vcc/gnd (Issue #163).
The 'global' method above is effectively an approximation of this where
it is assumed all pins in the architecture are tie-able.