docs: Use more human readable tag names.
Ids now use the name of the tag by default. If there is a collision
because the same tag name is used at multiple levels, it will instead
use the last title element in the form: `tag-last_title-tag_name`.

Example without collision:
 * PB Type - [`#tag-pb_type`](https://docs.verilogtorouting.org/en/latest/vpr/arch/reference.html#tag-pb_type)

Collision example:
 * Interconnect MUX - [`#tag-interconnect-mux`](https://docs.verilogtorouting.org/en/latest/vpr/arch/reference.html#tag-interconnect-mux)
 * Wire Segments MUX - [`#tag-wire-segments-mux`](https://docs.verilogtorouting.org/en/latest/vpr/arch/reference.html#tag-wire-segments-mux)

It's a better way to fix #280.