vpr: Update breadth-first/timing-driver router to more-efficiently explore non-config edges
Instead of checking all edges for configurability, only check those
which are known to be non-configurable based on edge partitioning.

This avoid iterating through the configurable edges when only non-config
edges need to be checked. Since the number of non-config edges is
typically far smaller than the number of configurable edges this should
improve run-time.