vpr: Fix RR graph CHAN capacitance calc. for bi-dir architectures
When calculating CHANX/CHANY capacitance on bi-dir architectures the
Cout of the previous segment is annotated on the node (potentially)
before the node it self is processed. The previous approach would
overwrite this value with the metal capacitance leading to a much
smaller (incorrect) node capacitance.

This fixes the issue by adding the metal capacitance (instead of
overwriting).