flow: Update vtr_bidir regression golden results
The modification to fly-weight RR node R/C values caused operations to
be re-ordered when calculating RR node C values on bi-dir architectures,
causing small fluctuations in the calculated C values.

The magnitude of these changes is very small, but they cause the router
to produce a different optimization result (likely because the small
changes cause effective 'ties' to break a different way). This caused some
benchmarks to exceed the previous QoR bounds.

This has been confirmed to be the case by loading a pre-fly-weight RR
graph and confirming that the current code produces the same
optimization result.