ace: Fix ACE to work with newer version of ABC
This includes fixing an assumption of sequential indices in ABC objects,
as well as the fact that ABC no longer has clock information.

The extract_clk_from_blif.py script now determines the clock name which
is passed to ACE for activity analysis.

Since ACE uses libabc to process the BLIF, the clock names now have to be
restored in the BLIF produced by ACE, similar to the process being done
after ABC.

Co-authored-by: kmurray <k.murray@utoronto.ca>