vtr_flow: Restore multi-clock latch information after ABC
New versions of ABC do not properly track the clock and edge sensitivity
of BLIF's .latch primitives. As a result this information must be restored
after ABC.

This is done with the restore_multiclock_latch_information.pl script,
which matches .latch output names between the pre and post ABC BLIFs and
to restore the clock and edge sensitivity information.

The script also writes out a list of clocks it found, which is used to
determine the clock name ACE should use during activity estimation.

Co-authored-by: Aaron Graham <aaron.graham@unb.ca>