build: Update MingW cross-compile for updated ABC
The updated ABC currently doesn't compile with MingW.

For now, just build VPR with MingW.

Fundamentally this is related to how ABC's included CMakeLists.txt
works. It just calls out to ABC's Makefile which doesn't understand that
it is being cross-compiled. In the future, those interested in making
this work can fix these limitations.