vpr: Fix handling of unswept undriven nets in netlist
An invalid Pin ID at the driver location is used to indicate an undriven
net in the netlist.

However, previously the netlist compression code would (incorrectly)
remove the driver pin since it was an invalid ID. This only occurs if
the netlist is not explicitly cleaned to remove undriven nets, which is
not how VPR is typically used.

This change ensures that the driver pin location is preserved, even if
it is an invalid ID.