vpr: Require absolute Fc to be a multiple of 2 with undir routing
Previously, VPR would interpret an absolute Fc on a unidir routing
architecture as referring to pairs of tracks (one in each direction).
This was inconsistent with the behaviour on bidir architectures
(where an absolute Fc corresponded to the absolute number of tracks),
and the traditional definition of Fc.

We now require that the Fc values on unidir routing to be multiples of
2, corresponding to the number of *tracks* to connect (rather than the
number of pairs).