libarchfpga: Improve error message for CB/SB lists.
Actually include the current length and expected length.

Before;
```
Error 1: arch.xml:127 CB or SB depopulation is too long. It should be (length) symbols for CBs and (length+1) symbols for SBs.
```

After;
```
Error 1: arch.xml:127 CB or SB depopulation is too long (50). It should be 49 symbols for CBs and 50 symbols for SBs.
```