vpr: Remove lb_type_rr_graph edges to/form external pins with Fc=0
This improves the accuracy of the packer's model of external routing
connectivity, ensuring it  does not try to use such pins to reach the
global routing network (to which they are not connected).

Note that this likely breaks packing designs which use such pins
(e.g. in conjunction with direct connections, like carry-chains), since
they are no longer connected to the external source pin.