vpr: Update intra-block router to consider multiple external sources/sinks
The intra-block router will now choose one (of potentially multiple
sources/sinks) when routing signals into/out of a cluster, based on
routing connectivity.

For instance, this means it will choose an external source/sink which
has a possible path to/from the internal source/sink.

This fixes some issues where the router would choose a source/sink for
which there was no possible path to connect.

Note that this change only ensures the router picks a valid external
source/sink for which there is a valid path within the intra-block RR
graph -- it does not guarantee that the 'right' external source/sink is
picked to ensure connections can be router between the blocks in the
iter-block RR graph. Making the correct choice remains future work.