flow: Add support for sweeping tool parameters to flow/task scripts
This allows a single task to sweep through a set of tool parameters with
the 'script_parameters_list_add=' option, similarly to the previous
'circuit_list_add=' and 'arch_list_add=' config file parameters.

Note that this adds an additional level to the run directory hierarchy
to uniquify the run directories. For example:

    run005/EArch.xml/styr.blif/common_--pack
    run005/EArch.xml/styr.blif/common_--pack_--place

Where the common parameters set with 'script_params=' or
'script_params_common=' are not shown, but the parameters added with
'script_params_list_add=' are added after 'common'. If no extra params
are added to the list the last level directory is just 'common'.