flow: Add support for 'failure' regression tests
The run_vtr_task.pl script now knows to look at the vpr_status field of
the golden results, and pass '-expect_fail' to run_vtr_flow.pl if the
golden result is a failure.

In run_vtr_flow.pl, when -expect_fail is set, it inverts the behaviour
when a tool fails (i.e. it reports 'OK*' and a zero exit code,
since the failure is expected).

This will allow the creation of regression tests which have expected
failures (e.g. to test error handling on invalid input).

Also refactors the golden result look-up functionality in
run_vtr_task.pl to be more general and remove code duplication.