vpr: Increase router congestion effort when approaching max routing iterations
If the router is approaching the maximum number of router iterations and
still has unresolved congestion it now switches into a more congestion
resolution focused mode. The hope is that this will improve overall
routability in otherwise marginal cases. It currently is activated when
the routing iteration number is >= 0.8 * max_router_iterations.

Currently this congestion-oriented mode does the following:
  1) Disables incremental re-routing of connections due to delay
     increases (i.e. it will not rip-up and re-route connections which
     are legal, but sub-optimal for delay).
  2) It increases the router bounding boxes for all nets to the entire
     device. This gives the router more flexibility to explore
     alternative paths around otherwise congested regions (at the cost
     of increased run-time).

On the VTR benchmarks this results in a 3% reduction in minW (~1% from
no delay based re-routing, and ~2% from removing bounding boxes), with no
CPD degradation.