vpr/pack: Support searching multiple modes.
This makes the packer work when there are multiple modes which change
the routing pattern through a cluster. The packer will try and pack
everything together, then it will fail routing causing it to fall back
to checking routing atom-by-atom. This means the packer will not add the
conflicting modes into one cluster.

This fixes the issue with "route through" modes having to be the
"zeroth" mode.

(I think this actually fixes #278 too.)