archfpga/vpr: Allow mixing <auto_layout> and <fixed_layout> gird specs
Previously <auto_layout> and <fixed_layouts> were exclusive. We now
allow zero-or-one <auto_layout> and zero-or-more <fixed_layouts> to be
specified.

If both an <auto_layout> and one-or-more <fixed_layout>s are specified
the <auto_layout> will be used by default, unless an explicit device
name is specified with VPR's --device option.