vpr: Add 'auto' option to --allow_unrelated_clustering and set as default
In auto mode VPR will automatically determine whether to enable unrelated
clustering based on the packing density achieved.

In the current implementation VPR attempts to pack the circuit first with
unrelated clustering disabled; as this results in more natural clusters
which are typically more routable (reduced minimum channel width).

If the resulting packing uses more resources than are available on the targeted
device, VPR will attempt a second packing with unrelated clustering enabled
in hope of fitting onto the target device.

Note that if the device is an auto-sized layout (instead of fixed layout)
this is equivalent to --allow_unrelated_custering off.

If there is a set of fixed layouts of varying sizes, VPR will pack
without unrelated clustering and only repack with unrelated clustering if
the design fails to fit on any of the fixed layouts.

If a specific fixed layout is specified (i.e. with --device) then VPR
will first attempt clustering with unrelated clustering off, and will
repack with unrelated clustering enabled only if the design did not fit
initially.