flow: Wrap run_vtr_task calls to run_vtr_flow in a script
This makes it easy to reproduce a particular task's flow invocation,
since the command used is not saved to disk (as vtr_flow.sh in the
generated run directory).

This also adds the '-system scripts' option to run_vtr_task which will
generate and print (but not execute) the associated tasks set of flow
scripts. This should make it easier to leverage the run_vtr_task
infrastructure to generate a set of flow runs, but allow users to handle
how they are actually run (e.g. locally, or on a compute cluster).
For more details see the updated run_vtr_task documentation.