vpr: Decrease incremental reroute net fanout threshold to 16
Previously, was set to 64 to avoid degrading critical path delay.
However, since the router will now automatically rip-up delay degraded
connections this value can be lowered.

On the VTR and (a subset of) the Titan benchmarks the new threshold showed
a 13% and 5% reduction in fixed channel width route-time respectively,
compared to the old default.