vpr: Incrementally increase router bounding boxes
When the router is in CONFLICTED mode (previously named CONGESTED mode)
it indicates there are some conflicting routing patterns which are not
resolving. This typically shows itself as a small number of overused RR
nodes which fail to resolve after many routing iterations.

Instead of completely blowing away the bounding boxes (which can hurt
run-time, particularly when performing minimum channel width search,
where unroutable solutions may be attempted), we now incrementally
increase the size of the bounding boxes. This increases the router's
search space and can allow signals to move out of the way of the routing
conflict (e.g. if they were previously at the edge of their bounding box).