vpr: Allow --analysis to force analysis of an illegal implementation
Starting with 6f57bee VPR would exit after a routing failure and would
not perform analysis. This prevents someone from naively parsing invalid
results (e.g. wirelength, critical path delay) from VPR's output if the
implementation was illegal (and they failed to check VPR's exit code or
for 'unrouteable' messages in the log file).

However there are cases where performing an analysis of an illegal
routing is useful (e.g. for debugging). Now, if --analysis is explicitly
specified the analysis stage (routing stats, final timing analysis, etc.)
are run *regardless* of whether the implementation is legal.

Use this with care as the resulting metrics are not valid, as the
implementation is illegal.