vpr: Improve router handling of high-fanout nets
This implements Jordan Swartz's high-fanout net routing optimizations.

We now maintain a grid-based spatial look-up for the nodes used by high
fanout nets. This look-up is used to quickly identify nodes used by the
current net which are spatially 'close' to current sink.

Instead of putting the entire existing route tree for the high fanout
net into the heap we only put the route tree nodes which are spatially
close to the sink.

This achieves two things:

  1) The spatial look-up avoids traversing the entire route tree to add
     it to the heap for each sink (since high fanout nets have large route
     trees, and many sinks this may have been expensive)
  2) By placing only partial routing nodes which are close to the sink
     onto the heap, the router will explore more localized area close to
     the sink.