vpr: Reduce memory usage during RR graph construction
Previously we dynamically created each edge in the RR graph which caused
irregular growth of each RR node's array of edges. This played havoc
with the memory allocator (fragmentation) resulting in a significant
increase in memory usage.

Instead, we now record all the edges to be created and allocate each node's
edges together. Since the edge arrays are sized exactly (and not resized
multiple times) this significantly reduces fragmentation and resulting
memory consumption.

For example on the Titan23 benchmark 'neuron' this results in a 50%
reduction in memory use (+6407.3 MiB to +3251.0 MiB during
build_rr_graph()). Since build_rr_graph() is the high-watermark of VPR's
memory use this results in ~35% reduction in peak memory use.