vpr: Maintain switchpoint order as specified in architecture file
This means that a statement like:

    <wireconn num_conns="3*to" from_type="L16" from_switchpoint="0,12,8,4" to_type="L16" to_switchpoint="0"/>

Inidcates an ordering of switchpoints.
When assigning wire connections first those from wires with switchpoint
0 are considered, then those with switchpoint 12, then 8, then 4.

In the above example, this is used to indicate a preference for driving
long length 16 wires from switchpoints near the *end* of other length 16
wires. (Switchpoint 0 indicates the end-point of a wire when used as a
from switchpoint).