vpr: Improve robustness of high-fanout routing on some architectures
On some architectures (e.g. those with more limited routing connectivity),
high-fanout routing could cause spurious failues (unable to find a
routing path). This is usually an artifact of high fanout routing,
caused by a poor choice of the route tree subset which is placed back on
the heap.

A typical spurious failure is for the high-fanout code to put only a single
routing wire back on the heap, which will be unable to reach the target
with the specific wirelengths and high-fanout connection-specific bounding box
(e.g. due to being in the wrong channel, or near the edge of the chip).

To improve robustness, if we detect a failure to route for a high-fanout
connection we try again using the full route tree (and issue a warning).
This is usually a rare occurrence, so the run-time penalty should be
minor.