vpr: Tweak high fanout routing to select more of the previous route tree
Previously, we would stop searching nearby spatial look-up bins if *any*
routing was found within the target bin. To improve robustness, we now
only stop if the target bin had more than 2 RR nodes within it.

This reduces the likely hood of ending up with only a single wire which
has limited flexibility and could cause the high-fanout connection
routing to fail.

Since we now retry if no path is found, this serves to reduce the number
of times we fallback on using the full route tree; it does not effect
whether a path can be found.