vpr: Add early exit heuristic for reconvergent routing
We now give-up when attempting to have routing converge multiple times,
if the CPD improvement after scaling back pres_fac is small (e.g. < 1%
of best legal routing CPD).

Typically the gain needs to be more in-order for there to be any CPD
improvement in a final legal routing (i.e. the low pres_fac routing
gives a pseudo-lower-bound on CPD). This avoids wasting CPU time trying
to get the router to re-converge for no benefit.