vpr: Do not absorb buffers which drive primary outputs
While we can absorb buffers which drive precisely one primary output,
doing so requires changing the name of the net to that of the primary
output. This make it difficult to map between net names in different
tools like VPR and Quartus. Since this is a relatively rare occurrence
simply not absorbing them should have minimal effect on resource usage.

However it ensures that nets effected by buffer absorbtion will maintain
the upstream net name. This makes it easier to, for instance, identify
clock names which can be used in both VPR and Quartus for the Titan
benchmarks.