vpr: Fix issue with eblif .conn statements delcared before net sinks
Previously VPR processed the .conn (merging the two associated nets) at
the point where the .conn was delcared.

This is incorrect in the case where the target net has not yet had all
it's sinks declared.

To handle such cases correctly we now record the pairs of nets which
need to be merged to resolve all .conn's and process them at the end of
the current .model. This ensures that all drivers/sinks have been associated
with their respective nets.

Closes #480