vpr/pack: rerouting when intra lib_nets have differing modes
This is to solve the issue for which there is a mode conflict and
the clusterer is not able to produce a valid cluster. The idea is to try
to re-route everytime there is a mode conflict between nets, avoiding
the usage of the modes that do result in a conflict.

Signed-off-by: Alessandro Comodi <acomodi@antmicro.com>