vpr: Fix expansion of large non-configurably connected node groups
Previously, we only expanded the lowest cost element of the
non-configurably connected group of RR nodes. Depending on the RR graph
structure this may mean that no path is found. To fix this we now expand
from all nodes in the non-configurably connected node set.

Also encapsulates single neighbour expansion in a function