Avoid duplicate work during flow.
These are two cases where work is being done twice.  The first case is
that the routing graph is being generated (or read from a file) twice.
The second is that when a routing graph load is done, the check_route
routine is called twice, once in the load function and another in the
root level flow function.

Signed-off-by: Keith Rothman <537074+litghost@users.noreply.github.com>