vpr: Invert swaps from single 'from' blocks to a macro 'to' blocks
Currently the placer's block swap code only supports:
 * Swapping two single blocks (neither part of macros)
 * Swapping a macro as 'from' block with a set of single 'to' blocks
Any other scenarios cause the swap to be aborted.

This had the unintended effect that a swap from a single 'from' block to
a 'to' block which is part of a macro was always aborted.

This commit adds support for this case by simply reversing the 'from'
and 'to' blocks, since the effect of the swap is the same regardless of
which block is actually the from/to.