vpr: Initial implementation of macro-macro placement swaps
Currently supports a subset of possible macro swaps where the
'to' macro does not extend outside the 'from' macro. This means that
swaps with partially overlapping macros remain unsupported.

We also only swap macros if they have the same shape, which is
sufficient for the common case of linear carry-chains.

Finally, self-swaps (i.e. shifts of the same macro) are also currently
unsupported.