vpr: Added support for using molecules with different pack_patterns within the same chain
This added code is the targeting the extra carry chain architecture
only where the logic block has two levels of carry chain one feeding
the other. This update handles the case were the second level of
addition in the netlist is longer the first level and spans more
logic blocks. The code now recognizes that and forces the molecules
to be placed in a way so that the chain is routable.