vpr: Updated class_size in cluster level pin feasibility check
Use the number of used pins in a class as the class_size if it is
already larger than the class_size multiplied by the external pin
utilization factor. This happens when the seed block chosen for
a cluster uses more inputs than the class_size multiplied by the
external pin utilization factor. Since when packing the seed block,
the external pin utilization factor is set to 1.0 to avoid large
molecules from failing packing.