odin: Fix a bug when comparing with the minimum hard multiplier size
When deciding if a multiplier should be implemented using soft logic
or the hard multiplier available in the architecture. We should compare
the size of the multiplier in the netlist to the minimum multiplier
size provided by the user instead of comparing the size of the hard
multiplier available in the architecture to the minimum multiplier size.