vpr: Making the search for transitive connectivity candidates more efficient.
Only look at nets that still have pins outside of the current
clusters. Since there is no point of looking at nets that are totally
absorbed into the cluster.

When searching for transitive candidates, search within a certain CLB
only once, since all candidates from this CLB will be added from
the first time.

Also, Instead of constraining the net size to explore, constrain the number
of CLBs to explore for each net. Since this is the time consuming part
of the code.