vpr: Add support for having multiple starting points along a chain pattern
Add support for defining multiple entry points for a chain pattern. This
means that in the architecture file at any point a long the chain a
mux could be added with one input driven by the Cout of the previous
Logic Block and the other input is connected to the Cin of the cluster.
Adding more freedom when placing small adders that are driven by global
constant nets (gnd/vcc)